Pages updated.

+++ src/pages/Auth/Login/Login.tsx
@@ -1,54 +1,37 @@
-import type { FunctionComponent } from 'react';
+import { type FunctionComponent, useState } from 'react';
 
 import { useNavigate, NavLink } from 'react-router-dom';
 
-import type { FormState } from '@/pages/Auth/_files/types';
 import type { LoginUserPayload } from '@/core/services/auth/types';
 import { AUTH_TOKEN_KEY } from '@/shared/files/consts';
-import { requiredEmail, requiredMin } from '@/shared/files/validations';
-import { clearFormMessage, validateForm } from '@/shared/helpers';
-import { useDispatch, useCustomState, useField } from '@/shared/hooks';
+import { useDispatch, useField } from '@/shared/hooks';
 import { loginUser } from '@/core/services';
 import { setUser } from '@/core/redux/reducers/auth';
-import { AlertDismissible, Loader, Input, Checkbox } from '@/shared/components';
+import { Loader, Input, Checkbox, Button } from '@/shared/components';
 import Auth from '../Auth';
 
-const initialState: FormState = {
-  loading: false,
-  formSubmitted: false,
-  serverErrors: {
-    email: [],
-    password: [],
-    request: [],
-  },
-};
-
 const Login: FunctionComponent = () => {
-  const [state, setState] = useCustomState<FormState>(initialState);
-  const { loading, formSubmitted, serverErrors } = state;
+  const [loading, setLoading] = useState(false);
 
   const dispatch = useDispatch();
   const navigate = useNavigate();
 
-  const email = useField({ name: 'email', validation: requiredEmail });
-  const password = useField({ name: 'password', validation: requiredMin(3) });
-  const keepLogged = useField<boolean>({ name: 'keep-logged' });
+  const email = useField();
+  const password = useField();
+  const keepLogged = useField<boolean>();
+
+  const handleChangeEmail: React.ChangeEventHandler<HTMLInputElement> = event => {
+    const { value } = event.target as HTMLInputElement;
+
+    email.ref.current.setCustomValidity(
+      value === 'demo@demo.com' ? 'This e-mail does not exists.' : '',
+    );
+  };
 
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
 
-    setState({ formSubmitted: true });
-
-    const isValidForm = validateForm([
-      { fields: [email], validation: requiredEmail },
-      { fields: [password], validation: requiredMin(3) },
-    ]);
-    if (!isValidForm) return;
-
-    setState({
-      loading: true,
-      serverErrors: { email: [], password: [], request: [] },
-    });
+    setLoading(true);
 
     try {
       const payload: LoginUserPayload = {
@@ -59,40 +42,14 @@ const Login: FunctionComponent = () => {
 
       const user = await loginUser(payload);
 
-      if (email.value === 'demo@demo.com') {
-        setState({
-          serverErrors: {
-            email: ['This e-mail does not exists.'],
-            password: ['The password is incorrect.'],
-            request: [],
-          },
-        });
-      } else {
-        if (keepLogged.value) {
-          localStorage.setItem(AUTH_TOKEN_KEY, user.token);
-        } else {
-          sessionStorage.setItem(AUTH_TOKEN_KEY, user.token);
-        }
+      if (keepLogged.value) localStorage.setItem(AUTH_TOKEN_KEY, user.token);
+      else sessionStorage.setItem(AUTH_TOKEN_KEY, user.token);
 
-        dispatch(
-          setUser({ firstName: user.firstName, lastName: user.lastName, email: user.email }),
-        );
-        navigate('/');
-      }
+      dispatch(setUser({ firstName: user.firstName, lastName: user.lastName, email: user.email }));
+      navigate('/');
     } catch (error: any) {
-      setState({
-        loading: false,
-        serverErrors: {
-          email: [],
-          password: [],
-          request: [error.message],
-        },
-      });
+      setLoading(false);
     }
-  };
-
-  const handleClearFormMessage = (field: string, index: number) => {
-    clearFormMessage(field, index, state, setState);
   };
 
   return (
@@ -103,56 +60,35 @@ const Login: FunctionComponent = () => {
         <h1 className="page-title">Login</h1>
 
         <div className="mb-3">
-          <Input type="email" label="E-mail address" field={email} formSubmitted={formSubmitted} />
-        </div>
-
-        {serverErrors.email.map((errorMessage, index) => (
-          <AlertDismissible
-            key={errorMessage}
-            variant="danger"
-            onDismiss={() => handleClearFormMessage('email', index)}
-          >
-            {errorMessage}
-          </AlertDismissible>
-        ))}
-
-        <div className="mb-3">
-          <Input type="password" label="Password" field={password} formSubmitted={formSubmitted} />
-        </div>
-
-        {serverErrors.password.map((errorMessage, index) => (
-          <AlertDismissible
-            key={errorMessage}
-            variant="danger"
-            onDismiss={() => handleClearFormMessage('password', index)}
-          >
-            {errorMessage}
-          </AlertDismissible>
-        ))}
-
-        <div className="form-check">
-          <Checkbox
-            label="Keep logged"
-            trueValue={true}
-            falseValue={false}
-            field={keepLogged}
-            formSubmitted={formSubmitted}
+          <Input
+            icon="Email"
+            label="E-mail"
+            type="email"
+            name="email"
+            required
+            placeholder="Enter your e-mail"
+            onChange={handleChangeEmail}
+            field={email}
           />
         </div>
 
-        {serverErrors.request.map((errorMessage, index) => (
-          <AlertDismissible
-            key={errorMessage}
-            variant="danger"
-            onDismiss={() => handleClearFormMessage('request', index)}
-          >
-            {errorMessage}
-          </AlertDismissible>
-        ))}
+        <div className="mb-3">
+          <Input
+            label="Password"
+            type="password"
+            name="password"
+            required
+            minLength={3}
+            placeholder="Enter your password"
+            field={password}
+          />
+        </div>
 
-        <button className="btn btn-primary d-block mx-auto" type="submit">
+        <Checkbox label="Keep logged" name="keep-logged" field={keepLogged} />
+
+        <Button className="d-block mx-auto" type="submit">
           Login
-        </button>
+        </Button>
 
         <div className="text-center">
           <hr className="mt-4" />
